Added the following scenarios to the login feature: 1. Login with correct credentials 2. Login with incorrect credentials
Added test steps to the two scenarios above as follows:
1. launch chrome browser
Test step to open a Chrome browser instance.

2. open login page
Test step to navigate to the saucedemo.com login page.

3. verify that user can log in
Test step to verify that a user is redirected to the home page after logging in using valid credentials (username and password).

4. verify that user cannot log in
Test step to verify that a user is not redirected to home page if they attempt to log in using invalid credentials.

5. close browser
Test step to close the browser instance.